Write tests for open, close, toggle
